hooks: use shellcheck

+++ hooks/ebuild_unpack_pre/tmpfs.bash
@@ -1,4 +1,4 @@
-#! /bin/bash
+#!/usr/bin/env bash
 
 hook_umount_tmpfs() {
 	local ignore_packages
@@ -14,7 +14,7 @@ hook_umount_tmpfs() {
 			elog "We are installing \"${package}\" so I'm unmounting $tmpfs"
 			umount "${tmpfs}"
 			# let's clean things, just in case
-			rm "${tmpfs}/*" -rf
+			rm "${tmpfs:?}/*" -rf
 			exit 0
 		fi
 	done
